+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Add Font::get_name().
 - Add image::RgbImage::convert().
 - Add ColorDepth::from_u8().
+- Add a widget_derive! macro for custom widgets, which basically implements Deref/DerefMut and adds convenience constructor and anchoring methods.
 
 ## [1.1.7] - 2021-08-04
 - Fix invalid cast in TextBuffer::selection_position().